requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-rest-registration==0.8.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
psycopg2==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
pydantic~=2.8.2
openai~=1.41.1
celery~=5.4.0